+++ doc/changes.rst
@@ -3,8 +3,8 @@ Change History
 
 .. currentmodule:: apsw
 
-next
-====
+3.43.1.0
+========
 
 All C code calling into Python and all C code called by Python uses
 vectorcall / fastcall (see :pep:`590`) which reduces the overhead of
@@ -21,8 +21,12 @@ passing and receiving positional and keyword arguments. (:issue:`477`
   from being 17% slower than pure SQLite to 2% slower.
 
 * A :source:`virtual table benchmark <tools/vtbench.py>` takes 35%
-  less time
+  less time.  (Remember that benchmarks are best case!)
 
+The :doc:`shell <shell>` JSON output modes have been fixed.  Mode
+'json' outputs a json array, while mode 'jsonl' does newline delimited
+json objects, aka `json lines <https://jsonlines.org/>`__.
+(:issue:`483`)
 
 3.43.0.0
 ========